Move styles to new files, and define some constants
