+++ src/Utils/colors.js
@@ -1,0 +1,9 @@
+const mainAppColor = '#53BCD0';
+const headerTintColor = '#FFF';
+const defaultTextColor = "#6F6F6F";
+
+export default {
+    mainAppColor,
+    headerTintColor,
+    defaultTextColor,
+};
