+++ src/Utils/navigation.js
@@ -1,0 +1,12 @@
+import { StyleSheet } from 'react-native';
+
+export const backImage = require('../assets/back-arrow.png');
+
+export const styles = StyleSheet.create({
+  myCustomHeaderBackImage: {
+    height: 21,
+    width: 27,
+    marginLeft: 12,
+    resizeMode: 'contain',
+  },
+});
