+++ src/components/IngredientsList/IngredientsList.js
@@ -1,4 +1,3 @@
-
 import React, { Component } from 'react';
 import { Text } from 'react-native';
