+++ src/screens/DrinkDetail/DrinkDetail.js
@@ -1,8 +1,9 @@
 import React, { Component } from 'react';
-import { View, ScrollView, Text, StyleSheet, Image } from 'react-native';
+import { View, ScrollView, Text, Image } from 'react-native';
 import { connect } from 'react-redux';
 import { getDrinkDetail } from '../../store/actions';
-import IngredientsList from '../../components/IngredientsList';
+import styles from './styles';
+import IngredientsList from '../../components/IngredientsList/IngredientsList';
 
 class DrinkDetailScreen extends Component {
 
@@ -46,62 +47,6 @@ class DrinkDetailScreen extends Component {
         }
     }
 }
-
-const styles = StyleSheet.create({
-    scrollView: {
-        backgroundColor: '#53BCD0',
-    },
-    detailContainer: {
-        flex: 1, 
-        flexDirection: 'row',
-        justifyContent: 'center',
-        width: "100%",
-        height: "100%",
-        backgroundColor: '#53BCD0',
-        padding: 20,
-    },
-    cardContainer: {
-        flex: 1, 
-        flexDirection: 'column',
-        backgroundColor: '#FFF',
-        padding: 20,
-        borderRadius: 5,
-        shadowColor: "#000000",
-        shadowOpacity: 0.5,
-        shadowRadius: 2,
-        shadowOffset: {
-          height: 1,
-          width: 1
-        }
-    },
-    imageContainer: {
-        flex: 1, 
-        flexDirection: 'column',
-        width: "100%",
-        height: 250,
-        marginBottom: 15,
-    },
-    image: {
-        width: "100%",
-        height: "100%",
-        borderRadius: 5
-    },
-    ingredient: {
-        fontSize: 15,
-        color: "#6F6F6F",
-        marginBottom: 4,
-    },
-    instructionsTitle: {
-        fontSize: 15,
-        color: "#6F6F6F",
-        marginTop: 10,
-        marginBottom: 10,
-    },
-    instructions: {
-        fontSize: 15,
-        color: "#6F6F6F",
-    }
-});
 
 const mapStateToProps = state => {
     return {
