+++ src/screens/Drinks/styles.js
@@ -1,0 +1,11 @@
+import { StyleSheet } from 'react-native';
+import colors from '../../Utils/colors';
+
+const styles = StyleSheet.create({
+    listContainer: {
+        width: "100%",
+        backgroundColor: colors.mainAppColor,
+    }
+});
+
+export default styles;
